Fixed app for Marshmallow (Mmm!) :metal:
Removed log code - all sorted now!
Added methods to check for permissions - Marshmallow now checks at
runtime.

During onCreate(), we'll check for permissions.

Also, using a lovely Toolbar to prompt the user about why we need to use
the permission, should they deny it originally. Also moved the
mViewPager member variable to the top so that I could access it in the
Toolbar.